added aws cli based deployment and code review feedback
