+++ .gitignore
@@ -1,0 +1,5 @@
+# ignore node_modules
+node_modules/
+
+# ignore generated zip files
+*.zip
