+++ scripts/S3BulkThumbnailCreator/package.json
