+++ scripts/S3ThumbnailCreator/package.json
